requirements.txt
altair==5.0.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
arviz==0.16.1
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cmdstanpy==1.1.0
colorama==0.4.6
comm==0.1.4
cons==0.4.6
contourpy==1.1.0
convertdate==2.4.0
cookiecutter==2.3.0
cycler==0.11.0
dacite==1.8.1
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
ephem==4.1.4
et-xmlfile==1.1.0
etuples==0.3.9
executing==1.2.0
fastjsonschema==2.18.0
fastprogress==1.0.3
filelock==3.12.2
fonttools==4.42.1
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
h5netcdf==1.2.0
h5py==3.9.0
holidays==0.31
htmlmin==0.1.12
humanize==4.8.0
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.14.0
ipyvue==1.9.2
ipyvuetify==1.8.10
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-execute-time==3.0.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
logical-unification==0.4.6
LunarCalendar==0.0.9
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
mdurl==0.1.2
miniKanren==1.0.3
mistune==3.0.1
multimethod==1.9.1
multipledispatch==1.0.0
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook_shim==0.2.3
numpy==1.23.5
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
pipreqsnb==0.2.4
platformdirs==3.10.0
plotly==5.16.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
prophet==1.1.4
protobuf==4.24.1
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.16.1
pymc==5.7.2
pymdown-extensions==10.1
PyMeeus==0.5.12
Pympler==1.0.1
pyparsing==3.0.9
pytensor==2.14.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
reacton==1.7.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rich-click==1.6.1
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
solara==1.19.0
soupsieve==2.4.1
# Editable install with no version control (src==0.1.0)
-e d:\git\the_iconic
stack-data==0.6.2
starlette==0.31.1
statsmodels==0.14.0
streamlit==1.25.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.23.2
validators==0.21.2
visions==0.7.5
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
widgetsnbextension==4.0.8
wordcloud==1.9.2
xarray==2023.8.0
xarray-einstats==0.6.0
yarg==0.1.9
ydata-profiling==4.5.1
zipp==3.16.2
